requirements.txt
datamol==0.11.4
dgl==1.1.2
einops==0.7.0
graphviz==0.20.1
networkx==3.1
pandas==2.0.3
Pillow==10.0.0
PyGCL==0.1.2
pygmtools==0.3.7
rdkit==2023.9.5
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
torch-cluster==1.6.1+pt20cpu
torch-geometric==2.3.1
torch-scatter==2.1.1+pt20cpu
torch-sparse==0.6.17+pt20cpu
torch-spline-conv==1.2.2+pt20cpu
torchviz==0.0.2
tqdm==4.65.0
wandb=0.16.4